backend/pyproject.toml
[project]
name = "super-deep-research"
version = "0.1.0"
description = "Multi-provider AI research application with intelligent report merging"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "openai>=1.88.0",
    "openai-agents>=0.0.19",
    "anthropic>=0.21.3",
    "httpx>=0.25.0",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.0",
    "reportlab>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0"
]
